feat: enhance billing process and user profile management (#202)
- Added duplicate subscription prevention in the billing service to avoid multiple active subscriptions for a user.
- Updated error handling in the BillingSettings component to provide user-friendly messages for duplicate subscriptions.
- Improved styling consistency in the LoginForm and RegisterForm components by ensuring background and text colors are uniform.
- Introduced a new ProfileForm component for user profile updates, including validation and formatting for phone numbers and other fields.
- Removed the deprecated Register page to streamline the registration process.

These changes enhance user experience by preventing duplicate subscriptions and improving the profile management interface.

+++ Frontend/src/components/settings/ProfileCard.tsx
@@ -20,11 +20,11 @@ const ProfileCard: React.FC<ProfileCardProps> = ({ user, onAvatarUpdate }) => {
   const getInitials = (firstName?: string, lastName?: string): string =>
     `${firstName?.charAt(0) || ""}${lastName?.charAt(0) || ""}`.toUpperCase();
 
-  // Format date helper
+  // Format date helper - shows first 3 letters of month
   const formatDate = (dateString?: string): string => {
     if (!dateString) return "N/A";
     const date = new Date(dateString);
-    return date.toLocaleDateString("en-US", { month: "long", year: "numeric" });
+    return date.toLocaleDateString("en-US", { month: "short", year: "numeric" });
   };
 
   // Format phone number for display
